requirements.txt
djangorestframework
drf_spectacular
drf_spectacular_sidecar
djangorestframework-simplejwt